requirements.txt
streamlit==1.41.1
gspread==6.1.4
google-auth==2.38.0
google-auth-oauthlib==1.2.1
anthropic==0.43.0
openai>=1.50.0
python-dotenv==1.0.1
pandas==2.2.3
